coffee_leaf_classifier_streamlit/coffee_leaf_classifier_streamlit/requirements.txt
streamlit==1.32.2
tensorflow-cpu==2.12.0
opencv-python-headless
matplotlib
Pillow
